requirements.txt
torch
torchvision
transformers
Pillow
timm
pandas
scikit-learn
tqdm